svdi-main/requirements.txt
pyvda
infi.systray
Pillow